Add files via upload
In this project, I have performed the job of a sports analyst. Here are given two datasets related to IPL (Indian Premier League) cricket matches. One dataset contains ball-by-ball data and the other contains match-wise data. I have imported the datasets into an SQL database and performed the tasks given in this assignment to find important insights from this dataset.
The first CSV file is for ball-by-ball data and it has information of all the 193468 balls bowled between the years 2008 and 2020. It has 17 columns.
The second file contains match-wise data and has data of 816 IPL matches. This table has 17 columns.